refactor: remove createTransport, move key-fetch to Identity.fetchFromServer
