+++ js/src/index.ts
@@ -9,7 +9,7 @@
 export { Identity } from './identity.js';
 export type { RequestContext, SessionRecoveryToken } from './identity.js';
 export { extractSessionRecoveryToken, decryptResponseWithToken, serializeSessionRecoveryToken, deserializeSessionRecoveryToken } from './identity.js';
-export { Transport, createTransport } from './client.js';
+export { Transport } from './client.js';
 export { PROTOCOL, HPKE_CONFIG } from './protocol.js';
 export {
   EhbpError,
